+++ Coding/Coding/1232_CheckStraightLine.cs
